creato la struttura HTML interamente (senza css e testo)
esercizio:
1. You can use HTML, JavaScript, and CSS to complete this project.
Plain CSS is recommended because that is what the lessons
have covered so far and you should get some practice with plain CSS

1. I can see a <main> element with a corresponding id="main-doc",
which contains the page's main content (technical documentation).

2. Within the #main-doc ( <main> ) element, I can see several <section> elements,
each with a class of "main-section". There should be a minimum of 5.

5. The .main-section elements should contain
at least 10 <p> elements total (not each).

6. The .main-section elements should contain
at least 5 <code> elements total (not each).

7. The .main-section elements should contain
at least 5 <li> items total (not each).

8. I can see a <nav> element with a corresponding id="navbar".

9. The navbar element should contain one <header> element
which contains text that describes
the topic of the technical documentation.

10. Additionally, the navbar should contain link (<a>) elements with the class of "nav-link".
There should be one for every element with the class "main-section".

11. The <header> element in the navbar
must come before any link (<a>) elements in the navbar.

12. Each element with the class of "nav-link" should contain text that corresponds to the <header> text within each <section>
(e.g. if you have a "Hello world" section/header,
your navbar should have an element
which contains the text "Hello world").

13. When I click on a navbar element, the page should navigate to the corresponding section of the "main-doc" element
(e.g. If I click on a "nav-link" element that contains the text "Hello world", the page navigates to a <section> element
that has that id and contains the corresponding <header>.